Simple visualization of stored data
First iteration of a GUI using PyQt
It shows a table with a row for each tracked game. The first Column is
the game's name, the last column is the price variation at the last
check performed. Columns inbetween show a history of previous records.

Added dummy database file for more convinient development/testing