List of tracked games from database instead of code
Now the scrapper gets the list of ids of the tracked products, instead
of having their urls in its code.
Those URLs were unneccesarily accesed every time the database was
updated. It is not the case since all the information needed comes from
the marketplace document, not the product page.